added loading screen, separated styles into directory, links working, tidied up console logs
